+++ src/main/java/firok/tiths/common/Potions.java
@@ -1,0 +1,10 @@
+package firok.tiths.common;
+
+import firok.tiths.potion.PotionHeavy;
+import net.minecraft.potion.Potion;
+
+// 状态效果的注册 reg_name是直接读取field名称
+public class Potions
+{
+	public static Potion heavy = new PotionHeavy(); // 沉重
+}