+++ src/main/java/firok/tiths/item/ItemRecord.java
@@ -1,0 +1,11 @@
+package firok.tiths.item;
+
+import net.minecraft.util.SoundEvent;
+
+public class ItemRecord extends net.minecraft.item.ItemRecord
+{
+	public ItemRecord(String name, SoundEvent sound)
+	{
+		super(name, sound);
+	}
+}